rm opencode provider, other misc updates
