+++ .env.example
@@ -2,4 +2,3 @@ AGENT_ONE_GOOGLE_GENERATIVE_AI_API_KEY=AIza****
 AGENT_ONE_GROQ_API_KEY=gsk_****
 AGENT_ONE_OPENROUTER_API_KEY=****
 AGENT_ONE_CEREBRAS_API_KEY=csk-****
-AGENT_ONE_OPENCODE_API_KEY=****
