+++ eslint.config.js
@@ -51,9 +51,6 @@ export default [
               "cerebras",
               "Cerebras",
               "Loadables",
-              "opencode",
-              "OpenCode",
-              "OPENCODE",
               "byok",
               "deeplink",
             ],
