+++ src/contexts/use-api-keys/api-keys-context.tsx
@@ -5,7 +5,6 @@ import {
   cerebrasApiKeyLoadableAtom,
   googleGenerativeAiApiKeyLoadableAtom,
   groqApiKeyLoadableAtom,
-  opencodeApiKeyLoadableAtom,
   openrouterApiKeyLoadableAtom,
 } from "@/lib/jotai/api-key-atoms";
 import { getLogger } from "@/lib/logger";
@@ -21,11 +20,10 @@ export const ApiKeysProvider: React.FC<{ children: ReactNode }> = ({
   const groq = useAtomValue(groqApiKeyLoadableAtom);
   const openrouter = useAtomValue(openrouterApiKeyLoadableAtom);
   const cerebras = useAtomValue(cerebrasApiKeyLoadableAtom);
-  const opencode = useAtomValue(opencodeApiKeyLoadableAtom);
 
   const loadables = useMemo(
-    () => [google, groq, openrouter, cerebras, opencode],
-    [google, groq, openrouter, cerebras, opencode],
+    () => [google, groq, openrouter, cerebras],
+    [google, groq, openrouter, cerebras],
   );
 
   const isApiKeysLoading = loadables.some((l) => l.state === "loading");
