+++ src/lib/jotai/provider-atoms.ts
@@ -1,20 +1,28 @@
 import { atomWithStorage } from "jotai/utils";
 
-import {
-  DEFAULT_PROVIDER_CONFIGS,
-  type ProviderConfig,
-  type ProviderId,
-} from "@/lib/settings/types";
-
 const SETTING_PREFIX = "agent-one-setting-";
+
+export type ProviderId = "cerebras" | "google" | "groq" | "openrouter";
+
+export interface ProviderConfig {
+  enabled: boolean;
+  headers: Record<string, string>;
+}
+
+const DEFAULT_PROVIDER_CONFIG: ProviderConfig = {
+  enabled: false,
+  headers: {},
+};
 
 const createProviderConfigAtom = (providerId: ProviderId) => {
   const key = `${SETTING_PREFIX}PROVIDER_CONFIG_${providerId.toUpperCase()}`;
   return atomWithStorage<ProviderConfig>(
     key,
-    DEFAULT_PROVIDER_CONFIGS[providerId],
+    DEFAULT_PROVIDER_CONFIG,
     undefined,
-    { getOnInit: true },
+    {
+      getOnInit: true,
+    },
   );
 };
 
@@ -22,12 +30,10 @@ export const cerebrasConfigAtom = createProviderConfigAtom("cerebras");
 export const googleConfigAtom = createProviderConfigAtom("google");
 export const groqConfigAtom = createProviderConfigAtom("groq");
 export const openrouterConfigAtom = createProviderConfigAtom("openrouter");
-export const opencodeConfigAtom = createProviderConfigAtom("opencode");
 
 export const providerConfigAtoms = {
   cerebras: cerebrasConfigAtom,
   google: googleConfigAtom,
   groq: groqConfigAtom,
   openrouter: openrouterConfigAtom,
-  opencode: opencodeConfigAtom,
 } as const;
