+++ src/routes/settings/sections/providers/provider-list-item.tsx
@@ -11,7 +11,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Switch } from "@/components/ui/switch";
-import type { ProviderConfig, ProviderId } from "@/lib/settings/types";
+import type { ProviderConfig, ProviderId } from "@/lib/jotai/provider-atoms";
 
 interface ProviderListItemProps {
   providerId: ProviderId;
